requirements.txt
numpy==1.15.4
python-mnist==0.6
pillow==5.3.0
